+++ src/components/Router/StepsRouter.js
@@ -20,10 +20,9 @@ class StepsRouter extends Component {
   render = ({options: {...globalUserOptions}, ...otherProps}) => {
     const componentBlob = this.currentComponent()
     const CurrentComponent = componentBlob.component
-    const i18n = this.props.options.i18n
     return (
       <div>
-        {!this.props.disableBackNavigation && <NavigationBar back={this.props.back} {...{i18n}} />}
+        {!this.props.disableBackNavigation && <NavigationBar back={this.props.back} i18n={this.props.i18n} />}
         <CurrentComponent
           {...{...componentBlob.step.options, ...globalUserOptions, ...otherProps}}
           trackScreen = {this.trackScreen}