+++ src/components/crossDevice/GenericError/index.js
@@ -12,7 +12,7 @@ class GenericError extends Component {
   render ({i18n}) {
     return (
       <div>
-        <Title title={i18n.t('cross_device.generic_client_error.message')} subTitle={i18n.t('cross_device.generic_client_error.instruction')} />
+        <Title title={i18n.t('errors.generic_client_error.message')} subTitle={i18n.t('errors.generic_client_error.instruction')} />
         <div className={theme.thickWrapper}>
           <span className={`${theme.icon}  ${style.icon}`} />
         </div>