+++ test/features/helpers/i18n_helper.rb
@@ -1,0 +1,12 @@
+class I18nHelper
+  def load_locale(locale)
+    tag = locale ? locale : 'en'
+    lines = File.readlines(File.expand_path("../locales/#{tag}.js"))
+    lines[0] = "{\n"
+    @translations = lines.join.to_json
+  end
+
+  def translate(key)
+    @translations[key]
+  end
+end